BUG: sparse: fix `idx_dtype` when building index arrays in many parts of sparse (#21824)
* fix idx_dtype treatment in random_array, block_diag and kron

* handle idx_dtype of indices in linalg.spsolve, min_max, coo.reshape, random_array, block_diag, kron, cs.multiply

* linting

* linting format strings

* fixes gh-20389: maintain idx_dtype across h/vstack. Also fixes block_diag

* fix f-string typos

+++ scipy/sparse/_data.py
@@ -187,7 +187,7 @@ class _minmax_mixin:
             value[not_full] = min_or_max(value[not_full], 0)
 
         mask = value != 0
-        major_index = np.compress(mask, major_index)
+        major_index = np.compress(mask, major_index).astype(idx_dtype, copy=False)
         value = np.compress(mask, value)
 
         if isinstance(self, sparray):
